post

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,34 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        // $tasks = Task::all();
         
-        return view("tasks.index", [
-            "tasks" => $tasks,
+        // return view("tasks.index", [
+        //     "tasks" => $tasks,
+        //     ]);
+        
+        $data = [];
+        if (\Auth::check()) {
+            
+            $tasks = Task::all();
+            
+            
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy("created_at", "desc")->paginate(10);
+            
+            
+            $data = [
+                "user" => $user,
+                "tasks" => $tasks,
+                ];
+            
+            return view("tasks.index", [
+                "tasks" => $tasks,
             ]);
+            
+        }
+        
+        return view("welcome", $data);
         
     }
 
@@ -30,11 +53,15 @@ class TasksController extends Controller
      */
     public function create()
     {
+        
+        $data = [];
+        
         $task = new Task;
         
         return view("tasks.create", [
             "task" => $task,
             ]);
+
     }
 
     /**
@@ -50,12 +77,14 @@ class TasksController extends Controller
             "status" => "required|max:10",
             ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            "content" => $request->content,
+            "status" => $request->status,
+            ]);
+        
         
         return redirect("/");
+        // return back();
     }
 
     /**
@@ -98,11 +127,13 @@ class TasksController extends Controller
     public function update(Request $request, $id)
     {
         $this->validate($request, [
+            "user_id" => "required|max:191",
             "content" => "required|max:191",
             "status" => "required|max:10",
             ]);
         
         $task = Task::find($id);
+        $task->user_id = $request->user_id;
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
@@ -118,8 +149,14 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
+        // $task = Task::find($id);
+        // $task->delete();
+        
+        if(\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
         
         return redirect("/");
     }
